#609 Simplified inlined CqActions
